trading accounts front end

+++ TradingAnalyzer.Core/Entities/Dtos/TradeDto.cs
@@ -24,40 +24,40 @@ namespace TradingAnalyzer.Entities.Dtos
         public DateTime EntryDate { get; set; }
 
         [DataType(DataType.Currency)]
-        public Double EntryPrice { get; set; }
+        public Decimal EntryPrice { get; set; }
         public TradingSetups EntrySetups { get; set; }
         public String EntryRemarks { get; set; }
         public String EntryScreenshot { get; set; }
 
         [DataType(DataType.Currency)]
-        public Double MFE { get; set; }
+        public Decimal MFE { get; set; }
 
         [DataType(DataType.Currency)]
-        public Double MFA { get; set; }
+        public Decimal MFA { get; set; }
 
         [DataType(DataType.DateTime)]
         public DateTime? ExitDate { get; set; }
 
         [DataType(DataType.Currency)]
-        public Double ExitPrice { get; set; }
+        public Decimal ExitPrice { get; set; }
         public TradeExitReasons ExitReason { get; set; }
         public String ExitRemarks { get; set; }
         public String ExitScreenshot { get; set; }
 
         [DataType(DataType.Currency)]
-        public Double Commissions { get; set; }
+        public Decimal Commissions { get; set; }
 
         public TradeClassifications Classification { get; set; }
 
         [DataType(DataType.Currency)]
         [Display(Name = "Profit/Loss")]
-        public Double ProfitLoss { get; set; }
+        public Decimal ProfitLoss { get; set; }
 
         public int Size { get; set; }
 
         [DataType(DataType.Currency)]
         [Display(Name = "Profit/Loss Per Contract")]
-        public Double ProfitLossPerContract { get; set; }
+        public Decimal ProfitLossPerContract { get; set; }
 
         public String TradingAccount { get; set; }
         public int TradingAccountId { get; set; }
